refactor(api): move CsvWithNoSessionDataError to certification/shared
